Add dependency injection to enable global HTTP retry policy
Following the [canonical suggestion of Microsoft](https://docs.microsoft.com/en-us/dotnet/architecture/microservices/implement-resilient-applications/use-httpclientfactory-to-implement-resilient-http-requests) to use dependency injection to create an IHttpClientFactory to both correctly use the [System.Net.Http](https://docs.microsoft.com/en-us/dotnet/api/system.net.http) and define a retry policy for network communication failures this implements basic dependency injection for the creation of the standalone "services" (which from a nomenclature standpoint appear to be their intent).

This is the MVP of a policy established to show that it compiles and breaks no tests (with only minor changes to constructor calls in the tests to allow DI to function). Additional tests will be needed once the appropriate retry policies have been selected.

ART-5892