Package dependency injection and unit test HTTP retry policy
Creating the service provider which uses an HTTP retry policy is an action that is separate from the execution of the CLI. This is a first step in packaging the application business logic into its own class which the CLI can interact with based on query parameters.

To test this application service provider a unit test was written that is poorly placed in the CcApiServiceTests but it was chosen as there is a no-response-content method there that will allow for the testing of the policy. While Polly is extensively tested and (recommends not testing that its internals behave correctly)[https://github.com/App-vNext/Polly/wiki/Unit-testing-with-Polly---with-examples#2-i-want-to-test-that-ive-configured-polly-policies-to-achieve-my-desired-resilience-strategy] this test establishes that we defined the resilience behavior as intended.