Add exception handling for API interactions
For operations against that API that rely on the content of the response in order to provide useful data we need to ensure that the status that is returned is in the 2xx range. If it is not we can use the logging framework to provide the user with actionable information.