Refactor CcApiService to diagnose template error
We are having consistent issues with customization of the non-TH and non-TB products. This existed on the API side and was likely resolved with cutchart-api commit  aefe84e3bcd238aa95daeb5459de1cf11a4a750a

Used exception handling rather than If-else statements to protect network and/or HTTP failures.

Updated the debugging command in launch.json to manifest the issue that we were experiencing.

TODO remove Systm.Console.Writeline() calls