Fix retry policy for API
Initially the request was retried if was not a bad request or had an authorization issue. This meant that successful requests (200s and 201s) were being retried 3 times needlessly. After those successive attempts the process was allowed to run to completion but it was terribly inefficient.

Now we have enumerated the responses that we expect the API to generate as copied from the spec as if these responses occur we should not retry as subsequent attempts will not function any differently. It is possible that this list is to extreme and only the OrTransientHttpError() and responses not in { OK, Created, Unauthorized, and Forbidden } should be retried but customer feedback of performance under load will help guide that.

Updated unit test to cover more of the policy retry types fr the ApiPolicy.

Removed status code-specific policies from the UpdateRetryPolicy as the possible set of responses from github are not well understood by the team.