Fix addresses used by the UpdateWithGithubAPI
The UpdateWithGithubAPI HttpClient base address was incorrectly targeting .../releases and lacked a trailing slash so it was being treated as a file rather than a folder. Thus when specifying paths to target with GetAsync() it was updating them in a haphazard way.

This restructures the way the URIs for the updater are made and resolves the construction error.