Add try/catch protection for all UpdateWithGithubAPI HttpClient calls
Similar to network access exception handling in other sections of the codebase the network calls were wrapped in try/catch blocks with EnsureSuccessStatusCode() validation on the responses prior to accessing the response content.

Additionally deep nested calls for reading a string and parsing it into JSON and accessing a specific field were broken into Linq-styled calls with the intent of readability.

Note: While we catch HttpRequestExceptions and top-level exceptions in the added try/catch blocks we are not providing new user messaging output as it is not clear what should be said. This mirrors previous code behavior as the calls, if they were not a successful status code, failed silently.