Handle HttpRequestExceptions separate of other operation exceptions
In the CcApiService we use EnsureSuccessStatusCode() to validate that the responses that are received from the cutchart-api service are correct before moving onto using their data. Previously failures like this were handled as top-level exceptions. By separating them out we are providing a developer hint that if they perform secondary operations on the response that could generate exceptions they may need to be handled differently. In this case we mark non-HTTP-request exceptions as unhandled and allow program execution.